--- a/.github/workflows/run-test-push.yml
+++ b/.github/workflows/run-test-push.yml
@@ -29,12 +29,12 @@ jobs:
           use-mamba: true
       - run: conda --version
       - run: python -V
-      - name: Install development version of bnlawrence/Pyfive:issue60
+      - name: Install development version of NCAS-CMS/Pyfive:h5netcdf
         run: |
           cd ..
-          git clone https://github.com/bnlawrence/pyfive.git
+          git clone https://github.com/NCAS-CMS/pyfive.git
           cd pyfive
-          git checkout issue60
+          git checkout h5netcdf
           pip install -e .
       - run: pip install -e .
       - run: conda list