--- a/.github/workflows/test_s3_remote_reductionist.yml
+++ b/.github/workflows/test_s3_remote_reductionist.yml
@@ -51,12 +51,12 @@ jobs:
           python-version: ${{ matrix.python-version }}
           miniforge-version: "latest"
           use-mamba: true
-      - name: Install development version of bnlawrence/Pyfive:issue60
+      - name: Install development version of NCAS-CMS/Pyfive:h5netcdf
         run: |
           cd ..
-          git clone https://github.com/bnlawrence/pyfive.git
+          git clone https://github.com/NCAS-CMS/pyfive.git
           cd pyfive
-          git checkout issue60
+          git checkout h5netcdf
           pip install -e .
       - name: Install PyActiveStorage
         run: |